Applied Revolution Smythes patch [#328] with some modifications. * Changed welcome message to same as in opensim * Inventory is only created to new users
Won't change the first and last name to regular names like in OpenSim. With realXtend style authentication it is important to know to which "domain" user belongs to.

Also newuser variable is not a class member because that might result a conflict if two or more users would try to login at same time.